requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.0
openai==1.68.0  # Used for OpenRouter API compatibility

# HTTP and API dependencies
httpx==0.28.1
httpcore==1.0.7
urllib3==2.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10

# Data handling and validation
pydantic==2.10.6
pydantic_core==2.27.2
numpy==2.2.4
typing_extensions==4.12.2
annotated-types==0.7.0

# Utilities
tqdm==4.67.1
jiter==0.9.0

# Optional dependencies
sounddevice==0.5.1  # Only if audio features are needed 